Fix unlink

+++ src/Controller/AdminCatController.php
@@ -82,7 +82,9 @@ class AdminCatController extends AbstractController
 
 
             if (!empty($_FILES['image']['name']) && empty($uploadedErrors)) {
-                unlink('uploads/' . $previousImage);
+                if (file_exists('uploads/' . $previousImage)) {
+                    unlink('uploads/' . $previousImage);
+                }
                 $fileName = uniqid() . '_' . $_FILES['image']['name'];
                 move_uploaded_file($_FILES['image']['tmp_name'], 'uploads/' . $fileName);
                 $cat['image'] = $fileName;
